feat: Refactor payment service interface and implementation structure
